6. make Company/Person/IBAN provider into Interface and change the concrete class into DefaultXXProvider 7. FairyModule .configure add implement for guice install-build 8. Fix some test cases after the interface modification 9. Zh locale use ZhCompanyProvider independently
